updated chat header
