+++ src/components/Channel/AddChannel.css
@@ -1,11 +1,13 @@
 .background {
-  width: 100%;
-  height: 100%;
+  width: 100vw;
+  height: 100vh;
   background: rgba(0, 0, 0, 0.8);
   position: fixed;
   display: flex;
   justify-content: center;
   align-items: center;
+  z-index: 99;
+  transform: translate(Y -60px);
 }
 
 .modal-wrapper {
@@ -16,8 +18,9 @@
   color: #000;
   display: grid;
   position: relative;
-  z-index: 10;
   border-radius: 10px;
+  transform: translate(Y -60px);
+  z-index: 99;
 }
 
 .modal-content {
@@ -52,37 +55,41 @@
   z-index: 10;
 }
 
-.form-container {
+.addch-form-container {
   display: flex;
   flex-direction: column;
 }
-input {
+.add-channel-input {
   width: 470px;
   height: 4vh;
   padding-left: 10px;
-}
-label {
-  width: 100%;
 }
 
 .sign-up-button {
   margin-top: 20px;
 }
-h1 {
-  padding-bottom: 15px;
-}
+
 .add-channel-content {
   padding-bottom: 15px;
   font-size: 18px;
   color: #141414;
 }
 
+.add-channel-header {
+  padding-bottom: 15px;
+}
 .name-label {
-  padding-bottom: 10px;
+  padding-bottom: 15px;
   font-size: 20px;
   font-weight: bold;
 }
 
-.form-container button {
-  width: 100px;
+.sidebar-button {
+  width: 280px;
+  background-color: #347a5a;
+  color: #fff;
+  border: 0;
+  height: 30px;
+  border-radius: 5px;
+  margin-top: 10px;
 }
