+++ src/components/Sidebar/Sidebar.js
@@ -18,8 +18,7 @@ import {
 } from "react-icons/ri";
 import SidebarOptions from "./SidebarOptions";
 import DirectMessages from "../DirectMessages/DirectMessages";
-import AddMembers from "../Channel/AddMembers";
-import ShowChannelMembers from "../Channel/ShowChannelMembers";
+
 
 function Sidebar(props) {
   const { userDb, recentDms, channelDb, setChat } = props;
@@ -34,17 +33,7 @@ function Sidebar(props) {
     setShowModal((prev) => !prev);
   };
 
-  //modal add members
-  const [showAddMembers, setShowAddMembers] = useState(false);
-  const openMemberModal = () => {
-    setShowAddMembers((prev) => !prev);
-  };
-
-  //modal show channel members
-  const [showMembers, setShowMembers] = useState(false);
-  const openAllMemberModal = () => {
-    setShowMembers((prev) => !prev);
-  };
+ 
 
   const displayChannels = channelDb
     ? channelDb.map((channel) => {
@@ -71,20 +60,11 @@ function Sidebar(props) {
         showModal={showModal}
         setShowModal={setShowModal}
       />
-      <AddMembers
-        onclick={openMemberModal}
-        showAddMembers={showAddMembers}
-        setShowAddMembers={setShowAddMembers}
-      />
-      <ShowChannelMembers
-        onclick={openAllMemberModal}
-        showMembers={showMembers}
-        setShowMembers={setShowMembers}
-      />
+     
       <div className="sidebar-container">
         <div className="sidebar-header">
           <div className="sidebar-info">
-            <h2>
+            <h2 className="sidebar-h2">
               Slack Clone{" "}
               <RiPencilFill
                 style={{
@@ -98,7 +78,7 @@ function Sidebar(props) {
               />
             </h2>
             <div className="user-name">
-              <h3>
+              <h3 className="sidebar-h3">
                 <RiCheckboxBlankCircleFill
                   style={{
                     marginTop: "1px",
@@ -114,20 +94,20 @@ function Sidebar(props) {
           </div>
 
           <div className="sidebar-options">
-            <button onClick={openModal}>Add Channel</button>
-            <button onClick={openMemberModal}>Add Members</button>
-            <button onClick={openAllMemberModal}>View Channel Members</button>
+        
             <SidebarOptions Icon={RiMenu2Line} title="All unreads" />
             <SidebarOptions Icon={RiChat1Line} title="Threads" />
             <SidebarOptions Icon={RiQuestionAnswerLine} title="All DMs" />
             <SidebarOptions Icon={RiAtLine} title="Mentions & reactions" />
             <SidebarOptions Icon={RiBookmarkLine} title="Saved Items" />
             <SidebarOptions Icon={RiMore2Fill} title="More" />
-            <SidebarOptions
-              onClick={openModal}
-              Icon={RiAddFill}
-              title="Add Channel"
-            />
+            <div onClick={openModal}>
+              <SidebarOptions
+                onClick={openModal}
+                Icon={RiAddFill}
+                title="Add Channel"
+              />
+            </div>
           </div>
         </div>
         <div
